// lib/types.ts
export type Role = "admin" | "commercial";
export type StatutCompte = "en_attente" | "actif" | "desactive" | "rejete";
export type StatutTransfert = "en_attente" | "transferee" | "en_livraison" | "livree" | "annulee";
export type ModePaiement = "especes" | "orange_money" | "mtn_momo" | "wave" | "autre";
export type CategorieProduct = "lunettes_photochromiques" | "lunettes_style" | "montres" | "accessoires";
export type TypeNotification = "info" | "succes" | "alerte" | "erreur";

export interface Profile {
  id: string;
  nom_complet: string;
  telephone: string;
  email?: string;
  ville: string;
  pays: string;
  role: Role;
  statut: StatutCompte;
  photo_url?: string;
  otp_verifie: boolean;
  created_at: string;
  updated_at: string;
}

export interface Produit {
  id: string;
  nom: string;
  reference: string;
  categorie: CategorieProduct;
  prix_base: number;
  photo_url?: string;
  qr_code?: string;
  description?: string;
  actif: boolean;
  created_by?: string;
  created_at: string;
  updated_at: string;
}
// ...
export interface TransfertCommande {
  id: string;
  commercial_expediteur_id: string;
  commercial_destinataire_id: string;
  vente_id?: string;
  client_telephone: string;
  client_ville: string;
  produit_id?: string;
  quantite: number;
  notes?: string;
  statut: StatutTransfert;
  created_at: string;
  updated_at: string;
  expediteur?: Profile;
  destinataire?: Profile;
  produit?: Produit;
}